Generic support added!
Although still in development, you can declare functions with generic types, see examples/generics.rsc
for more info. Will be squashing bugs for the next few days on this before merging.
Generics are costly as for each type passed into a function call a new instance of the function needs to be made
```x<int>() != x<string>()```
> However future support will optimise this out if the output of generic variations are equal.

I impld this mainly to aid in the std lib creation. A template method allows the std library
to accept any type and therefore be more versatile.